move generating commands under 'ksctl generate' subcommand
