+++ pkg/cmd/generate/cli_configs_test.go
@@ -1,4 +1,4 @@
-package adm
+package generate
 
 import (
 	"encoding/json"
