+++ pkg/cmd/generate/cluster_test.go
@@ -1,4 +1,4 @@
-package adm
+package generate
 
 import (
 	"fmt"
