+++ pkg/cmd/generate/permissions_test.go
@@ -1,4 +1,4 @@
-package adm
+package generate
 
 import (
 	"fmt"
