+++ pkg/cmd/generate/roles_manager.go
@@ -1,4 +1,4 @@
-package adm
+package generate
 
 import (
 	"fmt"
