+++ pkg/cmd/generate/roles_manager_test.go
@@ -1,4 +1,4 @@
-package adm
+package generate
 
 import (
 	"testing"
